+++ middleware/bufferparallel.go
@@ -5,7 +5,6 @@ import (
 	"strings"
 
 	"github.com/goyek/goyek/v3"
-	"github.com/goyek/goyek/v3/internal"
 )
 
 // BufferParallel is a middleware which buffers the output from parallel tasks
@@ -16,12 +15,12 @@ func BufferParallel(next goyek.Runner) goyek.Runner {
 			return next(in)
 		}
 
-		orginalOut := in.Output
+		originalOut := outputOrDiscard(in.Output)
 		streamWriter := &strings.Builder{}
-		in.Output = internal.SyncWriter(streamWriter)
+		in.Output = goyek.SyncWriter(streamWriter)
 
 		result := next(in)
-		io.Copy(orginalOut, strings.NewReader(streamWriter.String())) //nolint:errcheck // not checking errors when writing to output
+		io.WriteString(originalOut, streamWriter.String()) //nolint:errcheck // not checking errors when writing to output
 		return result
 	}
 }