+++ middleware/verbose.go
@@ -5,7 +5,6 @@ import (
 	"strings"
 
 	"github.com/goyek/goyek/v3"
-	"github.com/goyek/goyek/v3/internal"
 )
 
 // SilentNonFailed is a middleware which makes sure that only output from failed tasks is printed.
@@ -13,14 +12,14 @@ import (
 // The behavior is based on the Go test runner when it is executed without the -v flag.
 func SilentNonFailed(next goyek.Runner) goyek.Runner {
 	return func(in goyek.Input) goyek.Result {
-		orginalOut := in.Output
+		originalOut := outputOrDiscard(in.Output)
 		streamWriter := &strings.Builder{}
-		in.Output = internal.SyncWriter(streamWriter)
+		in.Output = goyek.SyncWriter(streamWriter)
 
 		result := next(in)
 
 		if result.Status == goyek.StatusFailed {
-			io.Copy(orginalOut, strings.NewReader(streamWriter.String())) //nolint:errcheck // not checking errors when writing to output
+			io.WriteString(originalOut, streamWriter.String()) //nolint:errcheck // not checking errors when writing to output
 		}
 
 		return result